More PR reviews fixes
